Fix GCC warnings.

+++ src/native/eventpipe/ep-json-file.c
@@ -112,7 +112,7 @@ ep_json_file_write_event_data (
 	int32_t characters_written = -1;
 
 	characters_written = ep_rt_utf8_string_snprintf (buffer, EP_ARRAY_SIZE (buffer), "{\"Time\" : \"%f\", \"Metric\" : \"1\",\n\"Stack\": [\n\"", millis_since_trace_start);
-	if (characters_written > 0 && characters_written < EP_ARRAY_SIZE (buffer))
+	if (characters_written > 0 && characters_written < (int32_t)EP_ARRAY_SIZE (buffer))
 		json_file_write_string (json_file, buffer);
 
 	if (message)
@@ -137,12 +137,12 @@ ep_json_file_write_event_data (
 		}
 
 		characters_written = ep_rt_utf8_string_snprintf (buffer, EP_ARRAY_SIZE (buffer), "\"%s!%s\",\n", assembly_name, method_name);
-		if (characters_written > 0 && characters_written < EP_ARRAY_SIZE (buffer))
+		if (characters_written > 0 && characters_written < (int32_t)EP_ARRAY_SIZE (buffer))
 			json_file_write_string (json_file, buffer);
 	}
 
 	characters_written = ep_rt_utf8_string_snprintf (buffer, EP_ARRAY_SIZE (buffer), "\"Thread (%" PRIu64 ")\"]},", ep_rt_thread_id_t_to_uint64_t (thread_id));
-	if (characters_written > 0 && characters_written < EP_ARRAY_SIZE (buffer))
+	if (characters_written > 0 && characters_written < (int32_t)EP_ARRAY_SIZE (buffer))
 		json_file_write_string (json_file, buffer);
 }
